pyproject.toml
# =====================================================================
# For a full TOML configuration guide, check the Flower docs:
# https://flower.ai/docs/framework/how-to-configure-pyproject-toml.html
# =====================================================================

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iot-fed"
version = "1.0.0"
description = ""
license = "Apache-2.0"
# Dependencies for your Flower App
dependencies = [
    "flwr[simulation]>=1.23.0",
    "flwr-datasets[vision]>=0.5.0",
    "torch==2.7.1",
    "torchvision==0.22.1",
    "timm>=0.9.0",
    "pillow>=10.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "minter"

# Point to your ServerApp and ClientApp objects
[tool.flwr.app.components]
serverapp = "iot_fed.server_app:app"
clientapp = "iot_fed.client_app:app"

# Custom config values accessible via `context.run_config`
[tool.flwr.app.config]
num-server-rounds = 5
fraction-train = 0.8
local-epochs = 2
lr = 0.0001
# Path to pretrained checkpoint (update this path to your actual checkpoint location)
checkpoint-path = "checkpoints/best_model.pth"
# Resume from previous federated learning result
resume-from-final = false  # true로 설정하면 final_model.pt에서 이어서 학습
# Dataset configuration
use-cifar10 = false  # Set to true to use CIFAR-10, false for custom skin disease dataset
data-root = "data/train"
val-data-root = "data/validation"

# Default federation to use when running the app
[tool.flwr.federations]
default = "remote-federation"  # 배포용

# Local simulation federation with 2 virtual SuperNodes (시우, 민지)
[tool.flwr.federations.local-simulation]
options.num-supernodes = 2

# Remote federation for 3 real devices (실제 배포용)
[tool.flwr.federations.remote-federation]
address = "127.0.0.1:9093"  # Control API (로컬에서 SuperLink에 연결)
insecure = true
options.num-supernodes = 2  # 시우, 민지 2명
